refactor(api-scheduler): split AWS-specific code into api-scheduler-aws package
The base api-scheduler package now contains only cloud-agnostic abstractions,
use cases, domain logic, and GraphQL schema. All AWS EventBridge-specific code
(EventBridgeSchedulerService, manifest loader, event handler, context wiring)
moves to the new api-scheduler-aws package. Integration tests follow the AWS
wiring and live in api-scheduler-aws.
