+++ packages/api-scheduler-aws/__tests__/__mocks/schedulerManifestPlugin.ts
