+++ packages/api-scheduler-aws/__tests__/features/useCases.test.ts
@@ -2,18 +2,18 @@ import { beforeEach, describe, expect, it } from "vitest";
 import type { CmsContext } from "@webiny/api-headless-cms/types/index.js";
 import { useHandler } from "~tests/__mocks/handler/useHandler.js";
 import { createMockScheduleClient } from "~tests/__mocks/scheduleClient.js";
-import { SchedulerService } from "~/shared/abstractions.js";
-import { VoidSchedulerService } from "~/features/SchedulerService/VoidSchedulerService.js";
-import { ScheduleActionUseCase } from "~/features/ScheduleAction/index.js";
-import { GetScheduledActionUseCase } from "~/features/GetScheduledAction/index.js";
-import { ListScheduledActionsUseCase } from "~/features/ListScheduledActions/index.js";
-import { CancelScheduledActionUseCase } from "~/features/CancelScheduledAction/index.js";
+import { SchedulerService } from "@webiny/api-scheduler/shared/abstractions.js";
+import { VoidSchedulerService } from "@webiny/api-scheduler/features/SchedulerService/VoidSchedulerService.js";
+import { ScheduleActionUseCase } from "@webiny/api-scheduler/features/ScheduleAction/index.js";
+import { GetScheduledActionUseCase } from "@webiny/api-scheduler/features/GetScheduledAction/index.js";
+import { ListScheduledActionsUseCase } from "@webiny/api-scheduler/features/ListScheduledActions/index.js";
+import { CancelScheduledActionUseCase } from "@webiny/api-scheduler/features/CancelScheduledAction/index.js";
 import {
     PublishTestEntryActionHandler,
     PublishTestEntryActionHandlerImpl
 } from "~tests/__mocks/PublishTestEntryActionHandler.js";
 import { NamespaceHandler } from "~tests/__mocks/NamespaceHandler.js";
-import { SCHEDULED_ACTION_PUBLISH } from "~/constants.js";
+import { SCHEDULED_ACTION_PUBLISH } from "@webiny/api-scheduler/constants.js";
 
 describe("Combined Use Cases", () => {
     let context: CmsContext;
