+++ packages/api-scheduler-aws/src/index.ts
@@ -1,0 +1,2 @@
+export { createScheduler } from "./createScheduler.js";
+export type { ICreateSchedulerParams } from "./createScheduler.js";
