+++ packages/api-scheduler/src/index.ts
@@ -7,4 +7,3 @@ export {
 export type { IScheduledAction } from "./shared/abstractions.js";
 
 export { ScheduledActionId } from "./domain/ScheduledActionId.js";
-export { createScheduler } from "./createScheduler.js";
