+++ packages/project-aws/_templates/appTemplates/api/graphql/package.json
@@ -30,6 +30,7 @@
     "@webiny/api-mailer": "latest",
     "@webiny/api-record-locking": "latest",
     "@webiny/api-scheduler": "latest",
+    "@webiny/api-scheduler-aws": "latest",
     "@webiny/api-website-builder": "latest",
     "@webiny/api-website-builder-scheduler": "latest",
     "@webiny/api-website-builder-workflows": "latest",
