+++ packages/project-aws/_templates/appTemplates/api/graphql/src/index.ts
@@ -23,7 +23,7 @@ import { createWebsockets } from "@webiny/api-websockets";
 import { registerWebsocketsDdbStorageOperations } from "@webiny/api-websockets-ddb";
 import { createRecordLocking } from "@webiny/api-record-locking";
 import { createHeadlessCmsScheduler } from "@webiny/api-headless-cms-scheduler";
-import { createScheduler } from "@webiny/api-scheduler";
+import { createScheduler } from "@webiny/api-scheduler-aws";
 import { createSchedulerClient } from "@webiny/aws-sdk/client-scheduler/index.js";
 import { createMailerContext, createMailerGraphQL } from "@webiny/api-mailer";
 import { createWorkflows } from "@webiny/api-workflows";
