+++ packages/project-aws/_templates/extensions/sqlite/api/graphql/src/index.ts
@@ -22,7 +22,7 @@ import { createWebsockets } from "@webiny/api-websockets";
 import { registerWebsocketsSqlStorageOperations } from "@webiny/api-websockets-sql";
 import { createRecordLocking } from "@webiny/api-record-locking";
 import { createSchedulerClient } from "@webiny/aws-sdk/client-scheduler/index.js";
-import { createScheduler } from "@webiny/api-scheduler";
+import { createScheduler } from "@webiny/api-scheduler-aws";
 import { createHeadlessCmsScheduler } from "@webiny/api-headless-cms-scheduler";
 import { createMailerContext, createMailerGraphQL } from "@webiny/api-mailer";
 import { createWorkflows } from "@webiny/api-workflows";
